Refactor touch indicator & remote UI code
