+++ app/Style.scss
@@ -639,11 +639,17 @@ span.line {
   .base_multi_touch_trigger {
     position: absolute;
 
-    &.touchpanel {
+    &.touch_indicator {
       pointer-events: none;
       top: 0;
       left: 0;
     }
+  }
+
+  .base_multi_touch_remote_affector {
+    position:absolute;
+    width: 100%;
+    height: 100%;
   }
 
   .base_picker {
@@ -772,6 +778,11 @@ span.line {
     background-image: url(images/leather.jpg);
     background-size: cover;
     color: $font_color_dark_primary;
+
+    .playerCounterContainer {
+      width: 100%;
+      height: 100%;
+    }
 
     .splashMultitouchInstruction {
       width: 90%;
