+++ app/components/base/Checkbox.tsx
@@ -2,10 +2,12 @@ import * as React from 'react'
 import FlatButton from 'material-ui/FlatButton'
 import CheckBoxIcon from 'material-ui/svg-icons/toggle/check-box'
 import CheckBoxOutlineIcon from 'material-ui/svg-icons/toggle/check-box-outline-blank'
+import RemoteRipple from './remote/RemoteRipple'
 
 export interface CheckboxProps {
   label: string;
   value: boolean;
+  remoteID?: string;
   onChange: (checked: boolean) => any;
 }
 
@@ -13,7 +15,7 @@ class ExpeditionCheckbox extends React.Component<CheckboxProps, {}> {
   render() {
     const icon = (this.props.value) ? <CheckBoxIcon/> : <CheckBoxOutlineIcon/>;
     return (
-      <span className="base_checkbox">
+      <RemoteRipple remoteID={this.props.remoteID} className="base_checkbox">
         <FlatButton onTouchTap={(e: any) => this.props.onChange(!this.props.value)}>
           <div>
             <span className="label">{this.props.label}</span>
@@ -21,7 +23,7 @@ class ExpeditionCheckbox extends React.Component<CheckboxProps, {}> {
           </div>
           <div className="subtext" id="subtext">{this.props.children}</div>
         </FlatButton>
-      </span>
+      </RemoteRipple>
     );
   }
 }
