--- app/components/base/RemoteTouchPanel.test.tsx
@@ -1,3 +1,0 @@
-describe('RemoteTouchPanel', () => {
-  // TODO
-});
