+++ app/components/base/remote/RemoteAffector.test.tsx
@@ -1,10 +1,4 @@
-
 describe('RemoteAffector', () => {
   it('Publishes interactions for remote clients');
   it('Publishes interaction with children that suppress events');
-});
-
-describe('RemoteRipple', () => {
-  it('Creates a ripple on inbound remote play event');
-  it('Publishes taps to remote clients');
 });
