+++ app/components/base/remote/RemoteRipple.test.tsx
@@ -1,0 +1,4 @@
+describe('RemoteRipple', () => {
+  it('Creates a ripple on inbound remote play event');
+  it('Publishes taps to remote clients');
+});
